requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-rest-knox==4.2.0
djangorestframework==3.13.1
Pillow==9.2.0
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
 django_qr_code==3.1.1